+++ include/tad/escalonador.hpp
@@ -8,13 +8,13 @@
 class Escalonador {
 private:
     MinHeap<Evento> m_fila_eventos;
-    double m_tempo_atual;
+    int m_tempo_atual;
     RedeLogistica* m_rede_logistica_ptr;
 
     int m_capacidade_transporte;
-    double m_latencia_transporte;
-    double m_intervalo_transportes;
-    double m_custo_remocao;
+    int m_latencia_transporte;
+    int m_intervalo_transportes;
+    int m_custo_remocao;
 
     int m_total_pacotes;
     int m_pacotes_entregues;
@@ -22,7 +22,7 @@ private:
     void processarEvento(const Evento& evento);
 
 public:
-    Escalonador(RedeLogistica* rede, int cap, double lat, double interv, double custo, int total_pacotes);
+    Escalonador(RedeLogistica* rede, int cap, int lat, int interv, int custo, int total_pacotes);
     void agendarEvento(const Evento& evento);
     void executar();
 };
